Isolate the GPU pool flatness test in its own binary
